% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_boxplot.R
\name{comparison_boxplot}
\alias{comparison_boxplot}
\title{Box plot of gene expression comparison between single-cell and bulk}
\usage{
comparison_boxplot(bulk, bulk_cell_type, ematrix, sc_cell_type, name, gene)
}
\arguments{
\item{bulk}{Expression bulk matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{bulk_cell_type}{A vector of cell types with each element representing
a cell type for the cell in the corresponding position in \code{bulk}}

\item{ematrix}{Expression single-cell matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{sc_cell_type}{A vector of cell types with each element representing a
cell type for the cell in the corresponding position in \code{ematrix}}

\item{name}{Title of the plot.}

\item{gene}{The marker gene going to be compared between single-cell and bulk}
}
\value{
the ggplot2 boxplot object
}
\description{
Box plot of gene expression comparison between single-cell and bulk
}
\examples{

}
