% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_heatmap.R
\name{DESeq2_heatmap}
\alias{DESeq2_heatmap}
\title{Heatmap of some seleted genes of bulk and single-cell expression.}
\usage{
DESeq2_heatmap(
  bulk,
  bulk_cell_type,
  ematrix,
  sc_cell_type,
  name,
  cell_list = c("DEC", "H1")
)
}
\arguments{
\item{bulk}{Expression bulk matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{bulk_cell_type}{A vector of cell types with each element representing
a cell type for the cell in the corresponding position in \code{bulk}}

\item{ematrix}{Expression single-cell matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{sc_cell_type}{A vector of cell types with each element representing a
cell type for the cell in the corresponding position in \code{ematrix}}

\item{name}{Title of the plot.}

\item{cell_list}{A two-element vector containing cell types which are waiting
for analysis using DESeq2}
}
\value{
a list containing the heatmap of bulk and single-cell
}
\description{
Heatmap of some seleted genes of bulk and single-cell expression. The
selected genes are Differential Expression Genes detected from bulk using
DESeq2
}
\examples{

}
