% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Four types of evaluation metrics}
\usage{
evaluation_metrics(ematrix, true_label, npc = 2)
}
\arguments{
\item{ematrix}{Expression matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{true_label}{True cell labels}

\item{npc}{Number of principal components used for dimension reduction}
}
\value{
a list containing adjusted Rand Index, Jaccard Index,
normalized mutual information, purity score, and silhouette
}
\description{
Four types of evaluation metrics
}
\examples{

}
\author{
Zi-Hang Wen \email{wenzihang0506@gmail.com}
}
