% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_analysis.R
\name{gg_analysis}
\alias{gg_analysis}
\title{The gene-gene interaction analysis using WGCNA and clusterProfile}
\usage{
gg_analysis(Bf.list.G, ont = "CC")
}
\arguments{
\item{Bf.list.G}{The list of gene latent matrices generated by Bfimpute}

\item{ont}{One of "BP", "MF", and "CC" subontologies. (from \code{enrichGO})}
}
\value{
a list of WGCNA network results and enrichGO results.
}
\description{
The gene-gene interaction analysis using WGCNA and clusterProfile
}
\examples{

}
