% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_pca.R
\name{plot_tsne_pca}
\alias{plot_tsne_pca}
\title{Title Scatter plot using t-SNE and first two PCs from PCA}
\usage{
plot_tsne_pca(res, name, cell_type)
}
\arguments{
\item{res}{counts Expression count matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{name}{Title of the plot.}

\item{cell_type}{Cell_type of the rows of \code{res}, aka the colour with which
you want to paint the dots.}
}
\value{
a list containing PCA plot and t-SNE plot (both are ggplot2 objects)
}
\description{
Title Scatter plot using t-SNE and first two PCs from PCA
}
\author{
Zi-Hang Wen \email{wenzihang0506@gmail.com}
}
