% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSCAN_trajectory.R
\name{TSCAN_trajectory}
\alias{TSCAN_trajectory}
\title{Evaluation using TSCAN}
\usage{
TSCAN_trajectory(ematrix, cell_type, name)
}
\arguments{
\item{ematrix}{Expression single-cell matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{cell_type}{A vector of cell types with each element representing a
cell type for the cell in the corresponding position in \code{ematrix}}

\item{name}{Title of the plot.}
}
\value{
a list containing the plot of trajectory, the plot of trajectory
pseudotime, POS score, and Kendall’s rank correlation score
}
\description{
Evaluation using TSCAN
}
\examples{

}
