% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Title Four types of evaluation metrics}
\usage{
evaluation_metrics(res, true_label, npc = 2)
}
\arguments{
\item{res}{counts Expression count matrix with rows corresponding to genes and
columns corresponding to cells.}

\item{true_label}{True cell labels}

\item{npc}{Number of principal components used for dimension reduction}
}
\value{
a list containing adjusted Rand Index, Jaccard Index,
normalized mutual information, purity score, and silhouette
}
\description{
Title Four types of evaluation metrics
}
\author{
Zi-Hang Wen \email{wenzihang0506@gmail.com}
}
